Changed long to int because I didn't realize long could be 8 bytes. This should fix byte ordering for network communications.

+++ updater/client.h
@@ -22,7 +22,7 @@ class client
 		
 		//unsorted public functions
 		void LoadDurationTable(const char *file);
-		void init_codepage(unsigned long arg);
+		void init_codepage(unsigned int arg);
 		
 		void register_char(int type);
 		
@@ -30,7 +30,7 @@ class client
 		sdl_user *graphics;
 		config *main_config;
 		connection *server;
-		unsigned long checksum;
+		unsigned int checksum;
 		pack *textpack;
 		pack *tilepack;
 		pack **spritepack;
@@ -45,7 +45,7 @@ class client
 		table todays_tip;
 		
 		int acp[2];	//-1, 0xec8a0 (a function pointer to badwords compare?)
-		unsigned long serverId;	//0
+		unsigned int serverId;	//0
 		unsigned char countryCode;
 		
 		int num_char_packs;
